init: add read_key() convenience
Currently the data link library allows users to define a simple
credential file where the api key is anticipated, but requires an
explicit call to read_key().  Introduce convenience logic to set the api
key in global configuration, if the file exist.

Update documentation in support of read_key() convenience.

Signed-off-by: Jamie Couture <jamie.couture@nasdaq.com>