api_config: standardize default config file location
Currently the config file path derivation logic is duplicated and
defines a default name in $HOME instead of tucking itself into a utility
directory.

Establish a default ~/.nasdaq directory for possible configuration files
and rename the apikey config file.  Adds logic to check if the file path
exists and add logic that tries hard to create the directories if they
don't exist.

Ideally the explicit call to read_key() should not be required by users
and the client should attempt to set they key for convenience.  A future
commit will attempt to add read_key() convenience.

Signed-off-by: Jamie Couture <jamie.couture@nasdaq.com>