connection: update session with proxies if available
It was erroneously assumed that new Session objects would include proxy
details, especially when env vars are defined.

Fixes an issue where users stuck behind a corporate firewall cannot
reach the api endpoint because the underlying liburl requests is not
updated.

Check if proxies are available and update the Session.

Signed-off-by: Jamie Couture <jamie.couture@nasdaq.com>